requirements.txt
torch~=1.5.0
numpy~=1.18.3
gym~=0.17.1
common~=0.1.2
pyzmq~=19.0.0